# src/main/java/com/example/onlinebookstore/domain/OrderItem.java
package com.example.onlinebookstore.domain;

import javax.persistence.*;
import java.math.BigDecimal;

@Entity
public class OrderItem {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;
    private String bookTitle;
    private BigDecimal price;
    private Integer quantity;

    public OrderItem() {}
    public OrderItem(String bookTitle, BigDecimal price, Integer quantity) { this.bookTitle = bookTitle; this.price = price; this.quantity = quantity; }

    public Long getId() { return id; }
    public void setId(Long id) { this.id = id; }
    public String getBookTitle() { return bookTitle; }
    public void setBookTitle(String bookTitle) { this.bookTitle = bookTitle; }
    public BigDecimal getPrice() { return price; }
    public void setPrice(BigDecimal price) { this.price = price; }
    public Integer getQuantity() { return quantity; }
    public void setQuantity(Integer quantity) { this.quantity = quantity; }
}